pyproject.toml
[project]
name = "git-clone-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "gitpython==3.1.43",
]
